backend/requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0

# RAG Pipeline
chromadb>=0.5.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0

# LLM Providers
google-generativeai>=0.7.0
ollama>=0.3.0

# Multimodal (CLIP)
open-clip-torch>=2.24.0
Pillow>=10.0.0

# Re-ranking
transformers>=4.40.0
torch>=2.0.0

# Evaluation
ragas>=0.1.0
datasets>=2.0.0

# Utilities
numpy>=1.26.0
pandas>=2.0.0
pyyaml>=6.0
httpx>=0.27.0
psutil>=5.9.0
